r/aws_sesv2_dedicated_ip_pool: remove dedicated_ips attribute
