+++ website/docs/r/sesv2_dedicated_ip_pool.html.markdown
@@ -35,13 +35,6 @@ The following arguments are optional:
 In addition to all arguments above, the following attributes are exported:
 
 * `arn` - ARN of the Dedicated IP Pool.
-* `dedicated_ips` - A list of objects describing the pool's dedicated IP's. See [`dedicated_ips`](#dedicated_ips).
-
-### dedicated_ips
-
-* `ip` - IPv4 address.
-* `warmup_percentage` - Indicates how complete the dedicated IP warm-up process is. When this value equals `1`, the address has completed the warm-up process and is ready for use.
-* `warmup_status` - The warm-up status of a dedicated IP address. Valid values: `IN_PROGRESS`, `DONE`.
 
 ## Timeouts
